python/requirements.txt
Flask==0.12
mysql-connector==2.1.4